backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.0
pdfplumber==0.10.3
google-generativeai==0.3.0
supabase==2.10.0
pymupdf==1.25.1
openai==1.55.0
PyMuPDF
weasyprint==62.3
pandas==2.2.0
resend==2.0.0
